geo-mapping-app monolithic/backend/requirements.txt
Flask
Flask-Cors
Flask-SQLAlchemy
Flask-Migrate
Flask-JWT-Extended
requests
geojson
stripe